Moving serial settings to a separate crate
