+++ Cargo.toml
@@ -31,7 +31,7 @@ features = []
 default-target = "thumbv7em-none-eabihf"
 
 [workspace]
-members = ["ad9959"]
+members = ["ad9959", "serial-settings"]
 
 [dependencies]
 embedded-storage = "0.3"
@@ -49,6 +49,7 @@ num_enum = { version = "0.7.1", default-features = false }
 paste = "1"
 idsp = "0.13"
 ad9959 = { path = "ad9959", version = "0.2.1" }
+serial-settings = {path = "serial-settings"}
 mcp230xx = "1.0"
 mutex-trait = "0.2"
 fugit = "0.3"
