+++ src/hardware/flash.rs
@@ -11,8 +11,14 @@ pub struct Settings {
     pub mac: smoltcp_nal::smoltcp::wire::EthernetAddress,
 }
 
+impl serial_settings::SerialSettings for Settings {
+    fn reset(&mut self) {
+        *self = Self::new(self.mac)
+    }
+}
+
 impl Settings {
-    fn new(mac: smoltcp_nal::smoltcp::wire::EthernetAddress) -> Self {
+    pub fn new(mac: smoltcp_nal::smoltcp::wire::EthernetAddress) -> Self {
         let mut id = heapless::String::new();
         write!(&mut id, "{mac}").unwrap();
 
@@ -22,9 +28,38 @@ impl Settings {
             mac,
         }
     }
+}
 
-    pub fn reset(&mut self) {
-        *self = Self::new(self.mac)
+pub struct Flash(pub LockedFlashBank);
+
+impl embedded_storage::nor_flash::ErrorType for Flash {
+    type Error = <LockedFlashBank as embedded_storage::nor_flash::ErrorType>::Error;
+}
+
+impl embedded_storage::nor_flash::ReadNorFlash for Flash {
+    const READ_SIZE: usize = LockedFlashBank::READ_SIZE;
+
+    fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error> {
+        self.0.read(offset, bytes)
+    }
+
+    fn capacity(&self) -> usize {
+        self.0.capacity()
+    }
+}
+
+impl embedded_storage::nor_flash::NorFlash for Flash {
+    const WRITE_SIZE: usize = stm32h7xx_hal::flash::UnlockedFlashBank::WRITE_SIZE;
+    const ERASE_SIZE: usize = stm32h7xx_hal::flash::UnlockedFlashBank::ERASE_SIZE;
+
+    fn erase(&mut self, from: u32, to: u32) -> Result<(), Self::Error> {
+        let mut bank = self.0.unlocked();
+        bank.erase(from, to)
+    }
+
+    fn write(&mut self, offset: u32, bytes: &[u8]) -> Result<(), Self::Error> {
+        let mut bank = self.0.unlocked();
+        bank.write(offset, bytes)
     }
 }
