+++ src/hardware/mod.rs
@@ -12,7 +12,6 @@ pub mod design_parameters;
 pub mod flash;
 pub mod input_stamper;
 pub mod pounder;
-pub mod serial_terminal;
 pub mod setup;
 pub mod shared_adc;
 pub mod signal_generator;
@@ -80,6 +79,8 @@ pub type SystemTimer = mono_clock::MonoClock<u32, MONOTONIC_FREQUENCY>;
 pub type I2c1 = hal::i2c::I2c<hal::stm32::I2C1>;
 pub type I2c1Proxy =
     shared_bus::I2cProxy<'static, shared_bus::AtomicCheckMutex<I2c1>>;
+
+pub type SerialTerminal = serial_settings::SerialTerminal<'static, UsbBus, flash::Settings, flash::Flash>;
 
 #[inline(never)]
 #[panic_handler]
